requirements.txt
asgiref==3.6.0
certifi==2023.5.7
charset-normalizer==3.1.0
coloredlogs==15.0.1
diffusers==0.12.1
Django==4.2.1
filelock==3.12.0
flatbuffers==23.5.9
fsspec==2023.5.0
huggingface-hub==0.14.1
humanfriendly==10.0
idna==3.4
importlib-metadata==6.6.0
invisible-watermark==0.1.5
Jinja2==3.1.2
MarkupSafe==2.1.2
mpmath==1.3.0
networkx==3.1
numpy==1.24.3
onnx==1.14.0
onnxruntime==1.15.0
opencv-python==4.7.0.72
packaging==23.1
Pillow==9.5.0
protobuf==4.23.1
psutil==5.9.5
PyWavelets==1.4.1
PyYAML==6.0
regex==2023.5.5
requests==2.31.0
sqlparse==0.4.4
sympy==1.12
tokenizers==0.12.1
torch==2.0.1
torchaudio==2.0.2
torchvision==0.15.2
tqdm==4.65.0
transformers @ git+https://github.com/huggingface/transformers.git@e42c86e6c9f884e5c8027d8b7056a1af8798c756
typing_extensions==4.6.1
urllib3==2.0.2
zipp==3.15.0
